update kill orphan.sh

+++ useful_scripts/kill-orphan.sh
@@ -11,18 +11,35 @@
 #ps -ef | grep '19561' | awk '{print $1, $2, $3}'
 #sudo kill -9 19561
 
+# note: you may check if your nvidia card is bind with invalid processes via:
+# pip install gpustat
+# gpustat
+# nvl  (or nvidia-smi)
+# when you see a gpu is in use in nvl (or nvidia-smi) but not int gpustat
+# you can use this script to kill orphan process
+
 main()
 {
     for pid in `sudo fuser /dev/nvidia*`; do
         #echo $pid
         out=`ps -ef | grep $pid`
-        user=`echo $out | awk '{print $1}'`
-        pid=`echo $out | awk '{print $2}'`
-        ppid=`echo $out | awk '{print $3}'`
 
-        if [ $ppid = '1' ]; then
-            echo "$user $pid $ppid"
-            hh=`sudo kill -9 $pid`
+        if [ ${#out} = 0 ]; then
+            echo "find non-exist process, pid=$pid"
+        else
+            user=`echo $out | awk '{print $1}'`
+            pid=`echo $out | awk '{print $2}'`
+            ppid=`echo $out | awk '{print $3}'`
+
+            #cat $out | while read line
+            #do
+            #    echo $line
+            #done
+
+            if [ $ppid = '1' ]; then
+                echo "find orphan process, user=$user pid=$pid ppid=$ppid"
+                hh=`sudo kill -9 $pid`
+            fi
         fi
     done
 }
